Extract admin module: mixin composition + lazy tool registration
Establishes the pattern for the book modularization refactor. Admin is
the first module peeled off the monolith; subsequent modules
(reconciliation, reporting, budgets, scheduling, investments, business)
will follow the same shape in their own PRs.

- Split book.py into a book/ package:
  - _base.py: BaseGnuCashBook + shared helpers (open, _resolve_guid,
    _find_account/_transaction/_split, serializers, _verify_*, exception)
  - _legacy.py: temporary home for every method not yet extracted;
    imports shared helpers from _base, inherits the rest via MRO
  - admin.py: AdminMixin (get/set/delete_account_slot)
  - __init__.py: build_book_class(modules) dynamically composes
    GnuCashBook from BaseGnuCashBook + LegacyMixin + enabled mixins,
    with a collision check across mixin method names
- Create a tools/ package for MCP tool wrappers:
  - _helpers.py: safe_tool + _json shared by every tool module
  - admin.py: register(mcp, get_book) that attaches admin tools
    (get_account_slots, set_account_slot, delete_account_slot,
    get_audit_log) on demand
- server.py: remove admin tool definitions and the local safe_tool/
  _json copies; _apply_module_filter lazy-imports extracted tool
  modules via importlib so their Pydantic schemas never build when
  the module is disabled; main() sets _book_class via build_book_class
  to match enabled modules
- test_modules.py: weaken test_all_registered_tools_are_mapped to a
  subset check (extracted tools lazy-load), add
  test_non_extracted_modules_are_fully_registered_at_import for the
  remaining invariant, and add TestExtractedModuleLazyLoading with
  4 new tests covering the lazy-load contract
- Backward compatible: top-level gnucash_mcp.book still exports
  GnuCashBook, GnuCashLockError, and the _verify_* helpers, so
  existing test imports and consumers keep working

Full suite: 602 passed (up from 587 — 15 new lazy-loading tests).
Live test confirmed tools/admin lazy-import fires correctly via
--modules=reporting,reconciliation,admin; get_account_slots and
get_audit_log both callable end-to-end.
