+++ src/gnucash_mcp/book/_legacy.py
@@ -1,519 +1,54 @@
-"""GnuCash book wrapper using piecash."""
+"""LegacyMixin — methods not yet extracted into per-module mixins.
+
+Temporary home for GnuCashBook methods that haven't been peeled off
+into their own module yet (core, reporting, budgets, scheduling,
+investments, business, reconciliation). As each module is extracted,
+methods move from here into their own file. When empty, this file
+can be deleted.
+
+Shared helpers (open, finders, serializers) live in _base.py.
+"""
 
 import logging
-import sqlite3
-import time
-from contextlib import contextmanager
 from datetime import date, datetime, timedelta
 from decimal import Decimal, InvalidOperation
-from pathlib import Path
-from typing import Generator
 
 import piecash
+
+from gnucash_mcp.book._base import (
+    _account_to_compact_line,
+    _account_to_dict,
+    _commodity_to_compact_line,
+    _lot_to_compact_line,
+    _split_to_dict,
+    _sx_to_compact_line,
+    _to_date,
+    _transaction_to_compact_line,
+    _transaction_to_dict,
+    _unreconciled_split_to_compact_line,
+    _upcoming_to_compact_line,
+    _verify_composite_write,
+    _verify_delete,
+    _verify_write,
+)
 
 # Debug logger - configured by logging_config.setup_logging()
 debug_logger = logging.getLogger("gnucash_mcp.debug")
 
 
-def _to_date(dt: date | datetime) -> date:
-    """Convert a datetime or date to a date object.
+# Module-level helpers, serializers, and GnuCashLockError all moved to _base.py
+# (imported at top of this file).
 
-    piecash may return either datetime or date for price dates depending
-    on how the price was created. This normalizes to date.
+
+class LegacyMixin:
+    """Methods not yet extracted into per-module mixins.
+
+    Depends on shared helpers from BaseGnuCashBook (via MRO):
+      - self.open, self.book_path
+      - self._find_account, self._find_transaction, self._find_split
+      - self._resolve_guid, self._require_default_currency
+      - self._get_or_create_currency, self._collect_descendants
     """
-    if isinstance(dt, datetime):
-        return dt.date()
-    return dt
-
-
-def _verify_write(session, table, guid: str, label: str) -> None:
-    """Verify a raw SQL INSERT persisted by reading back the primary key.
-
-    Must be called within the same session, before book.save().
-    Raises RuntimeError if the inserted row cannot be found.
-    """
-    from sqlalchemy import select, func
-
-    count = session.execute(
-        select(func.count()).select_from(table).where(table.c.guid == guid)
-    ).scalar()
-    if count != 1:
-        debug_logger.error(
-            f"Write verification FAILED: {label} guid={guid} count={count}"
-        )
-        raise RuntimeError(
-            f"Write verification failed: {label} with guid "
-            f"{guid} not found after INSERT (count={count})"
-        )
-
-
-def _verify_composite_write(
-    session, table, conditions: dict, label: str
-) -> None:
-    """Verify a raw SQL INSERT for a table with composite primary key.
-
-    Must be called within the same session, before book.save().
-    Raises RuntimeError if the inserted row cannot be found.
-    """
-    from sqlalchemy import select, func, and_
-
-    where_clause = and_(
-        *(table.c[col] == val for col, val in conditions.items())
-    )
-    count = session.execute(
-        select(func.count()).select_from(table).where(where_clause)
-    ).scalar()
-    if count != 1:
-        debug_logger.error(
-            f"Write verification FAILED: {label} conditions={conditions} "
-            f"count={count}"
-        )
-        raise RuntimeError(
-            f"Write verification failed: {label} not found "
-            f"after INSERT (count={count})"
-        )
-
-
-def _verify_delete(session, conditions: dict, label: str) -> None:
-    """Verify a raw SQL DELETE removed the expected row(s).
-
-    Must be called within the same session, before book.save().
-    Raises RuntimeError if matching rows still exist.
-    """
-    from sqlalchemy import text
-
-    where_parts = " AND ".join(f"{col} = :{col}" for col in conditions)
-    count = session.execute(
-        text(f"SELECT count(*) FROM slots WHERE {where_parts}"),
-        conditions,
-    ).scalar()
-    if count != 0:
-        debug_logger.error(
-            f"Delete verification FAILED: {label} still has {count} rows"
-        )
-        raise RuntimeError(
-            f"Delete verification failed: {label} still exists "
-            f"after DELETE ({count} rows remain)"
-        )
-
-
-class GnuCashLockError(Exception):
-    """Raised when the GnuCash book is locked by another process."""
-
-    pass
-
-
-def _account_to_dict(account: piecash.Account) -> dict:
-    """Convert a piecash Account to a serializable dict."""
-    return {
-        "guid": account.guid,
-        "name": account.name,
-        "fullname": account.fullname,
-        "type": account.type,
-        "commodity": account.commodity.mnemonic if account.commodity else None,
-        "description": account.description or "",
-        "placeholder": bool(account.placeholder),
-    }
-
-
-# Mapping of top-level parent to "obvious" account types that need no annotation
-_DEFAULT_TYPES = {
-    "Assets": {"ASSET"},
-    "Liabilities": {"LIABILITY"},
-    "Income": {"INCOME"},
-    "Expenses": {"EXPENSE"},
-    "Equity": {"EQUITY"},
-}
-
-
-def _account_to_compact_line(account: piecash.Account) -> str:
-    """Convert a piecash Account to a compact one-line string.
-
-    Format: "fullname [ANNOTATION]" where annotation is shown only when
-    the account type is non-obvious or the account is a placeholder.
-
-    Examples:
-        "Assets:Checking [BANK]"
-        "Expenses:Groceries [PLACEHOLDER]"
-        "Expenses:Groceries:Bakery"
-        "Assets:Investments [ASSET, PLACEHOLDER]"
-    """
-    fullname = account.fullname
-    annotations = []
-
-    top_level = fullname.split(":")[0]
-    default_types = _DEFAULT_TYPES.get(top_level, set())
-
-    if account.type not in default_types:
-        annotations.append(account.type)
-    if account.placeholder:
-        annotations.append("PLACEHOLDER")
-
-    if annotations:
-        return f"{fullname} [{', '.join(annotations)}]"
-    return fullname
-
-
-def _split_to_dict(split: piecash.Split) -> dict:
-    """Convert a piecash Split to a serializable dict."""
-    # Treat epoch zero (1970-01-01) as null — GnuCash stores this
-    # as the default for unreconciled splits instead of NULL
-    rec_date = split.reconcile_date
-    if rec_date and rec_date.year <= 1970:
-        rec_date = None
-    return {
-        "guid": split.guid,
-        "account": split.account.fullname,
-        "value": str(split.value),
-        "quantity": str(split.quantity),
-        "memo": split.memo or "",
-        "reconcile_state": split.reconcile_state,
-        "reconcile_date": rec_date.isoformat() if rec_date else None,
-        "lot_guid": split.lot.guid if split.lot else None,
-    }
-
-
-def _transaction_to_dict(transaction: piecash.Transaction) -> dict:
-    """Convert a piecash Transaction to a serializable dict."""
-    result = {
-        "guid": transaction.guid,
-        "date": transaction.post_date.isoformat(),
-        "description": transaction.description,
-        "currency": transaction.currency.mnemonic,
-        "splits": [_split_to_dict(s) for s in transaction.splits],
-    }
-    if transaction.notes:
-        result["notes"] = transaction.notes
-    return result
-
-
-def _commodity_to_compact_line(namespace: str, entry: dict) -> str:
-    """Convert a commodity dict to a compact one-line tab-separated string.
-
-    Format: "NAMESPACE:MNEMONIC\\tfullname\\tprice_info"
-
-    Examples:
-        "CURRENCY:USD\\tUS Dollar"
-        "FUND:VTSAX\\tVanguard Total Stock Market\\t128.75 USD (2026-02-07)"
-    """
-    prefix = f"{namespace}:{entry['mnemonic']}"
-    name = entry.get("fullname", "")
-    parts = [prefix, name]
-    lp = entry.get("latest_price")
-    if lp:
-        parts.append(f"{lp['value']} {lp['currency']} ({lp['date']})")
-    return "\t".join(parts)
-
-
-def _unreconciled_split_to_compact_line(split_dict: dict) -> str:
-    """Convert an unreconciled split dict to a compact one-line tab-separated string.
-
-    Format: "short_guid\\tYYYY-MM-DD\\tdescription\\tamount\\tstate"
-
-    The account name is omitted — the caller already knows which account
-    was queried. State is 'n' (new) or 'c' (cleared).
-
-    Examples:
-        "a1b2c3d4\\t2026-01-15\\tSafeway\\t-47.50\\tn"
-        "e5f6a7b8\\t2026-01-20\\tPayroll\\t3200.00\\tc"
-    """
-    short_guid = split_dict["guid"][:8]
-    d = split_dict["date"]
-    desc = split_dict["description"]
-    amount = split_dict["amount"]
-    state = split_dict["reconcile_state"]
-    return f"{short_guid}\t{d}\t{desc}\t{amount}\t{state}"
-
-
-def _transaction_to_compact_line(
-    transaction: piecash.Transaction, exclude_account: str | None = None,
-) -> str:
-    """Convert a piecash Transaction to a compact one-line tab-separated string.
-
-    Format: "YYYY-MM-DD\\tabcd1234\\tDescription\\tAccount amount, Account amount"
-
-    Fields are tab-separated so descriptions can be any length without
-    truncation — important for matching. Tabs are single-token for LLMs
-    and never appear in transaction descriptions.
-
-    Args:
-        transaction: piecash Transaction object.
-        exclude_account: If set, omit the split for this account (used when
-                        listing transactions filtered by account — the AI
-                        already knows the filtered account).
-
-    Examples:
-        "2026-02-15\\ta1b2c3d4\\tSafeway\\tExpenses:Groceries 47.50, Liabilities:Visa -47.50"
-        "2026-02-14\\te5f6a7b8\\tMonthly Rent\\tExpenses:Rent 1850.00, Assets:Checking -1850.00"
-    """
-    date_str = transaction.post_date.isoformat()
-    short_guid = transaction.guid[:8]
-    desc = transaction.description
-
-    parts = []
-    for split in transaction.splits:
-        account_name = split.account.fullname
-        if exclude_account and account_name == exclude_account:
-            continue
-        amount = split.quantity
-        # For multi-currency, show both quantity and value
-        if split.quantity != split.value:
-            currency = transaction.currency.mnemonic
-            commodity = split.account.commodity.mnemonic
-            parts.append(f"{account_name} {amount} {commodity} (={split.value} {currency})")
-        else:
-            parts.append(f"{account_name} {amount}")
-
-    splits_str = ", ".join(parts)
-    line = f"{date_str}\t{short_guid}\t{desc}\t{splits_str}"
-    if transaction.notes:
-        line += f"\t{transaction.notes}"
-    return line
-
-
-def _lot_to_compact_line(lot_dict: dict) -> str:
-    """Convert a lot dict to a compact one-line tab-separated string.
-
-    Format: "short_guid\\ttitle\\tquantity shares\\tcost_basis basis\\tclosed"
-
-    Examples:
-        "a1b2c3d4\\tVTSAX 2026-01-15 purchase\\t10.0000 shares\\t2504.50 basis"
-        "e5f6a7b8\\tAAPL 2025-06-01 purchase\\t0 shares\\t0 basis\\tCLOSED"
-    """
-    short_guid = lot_dict["guid"][:8]
-    title = lot_dict["title"]
-    qty = lot_dict["quantity"]
-    basis = lot_dict["cost_basis"]
-    parts = [short_guid, title, f"{qty} shares", f"{basis} basis"]
-    if lot_dict.get("is_closed"):
-        parts.append("CLOSED")
-    return "\t".join(parts)
-
-
-def _sx_to_compact_line(sx_dict: dict) -> str:
-    """Convert a scheduled transaction dict to a compact one-line tab-separated string.
-
-    Format: "short_guid\\tname\\tfrequency\\tnext:YYYY-MM-DD"
-
-    Examples:
-        "a1b2c3d4\\tMonthly Rent\\tmonthly\\tnext:2026-03-01"
-        "e5f6a7b8\\tWeekly Groceries\\tweekly\\tnext:2026-02-22"
-        "c9d0e1f2\\tOld Subscription\\tmonthly\\tdisabled"
-    """
-    short_guid = sx_dict["guid"][:8]
-    name = sx_dict["name"]
-    freq = sx_dict["frequency"]
-    if not sx_dict.get("enabled"):
-        status = "disabled"
-    elif sx_dict.get("next_occurrence"):
-        status = f"next:{sx_dict['next_occurrence']}"
-    else:
-        status = "no upcoming"
-    return f"{short_guid}\t{name}\t{freq}\t{status}"
-
-
-def _upcoming_to_compact_line(entry: dict) -> str:
-    """Convert an upcoming transaction dict to a compact one-line tab-separated string.
-
-    Format: "short_guid\\tname\\tYYYY-MM-DD\\tN days\\tamount"
-
-    Examples:
-        "a390ebb7\\tComcast Xfinity\\t2026-02-27\\t12 days\\t149.26"
-        "b1c2d3e4\\tMonthly Rent\\t2026-03-01\\t14 days\\t1850.00"
-    """
-    short_guid = entry["guid"][:8]
-    name = entry["name"]
-    occ_date = entry["occurrence_date"]
-    days = entry["days_until"]
-    amount = entry["amount"]
-    return f"{short_guid}\t{name}\t{occ_date}\t{days} days\t{amount}"
-
-
-class GnuCashBook:
-    """Thread-safe wrapper for piecash book operations."""
-
-    def __init__(self, book_path: str):
-        """Initialize with path to GnuCash SQLite book.
-
-        Args:
-            book_path: Path to the GnuCash SQLite file.
-
-        Raises:
-            FileNotFoundError: If the book path doesn't exist.
-            ValueError: If the file is not a valid SQLite GnuCash book.
-        """
-        self.book_path = Path(book_path)
-        if not self.book_path.exists():
-            raise FileNotFoundError(f"GnuCash book not found: {book_path}")
-
-    # Tables that support GUID resolution
-    _GUID_TABLES = frozenset({
-        "transactions", "splits", "accounts", "lots",
-        "schedxactions", "commodities", "budgets",
-        "customers", "vendors", "invoices",
-    })
-
-    def _resolve_guid(self, table: str, partial: str) -> str:
-        """Resolve a partial GUID prefix to a full 32-character GUID.
-
-        Uses raw SQLite in read-only mode — no piecash session needed.
-
-        Args:
-            table: Database table name (e.g., "transactions", "splits").
-            partial: Full or partial GUID prefix (minimum 8 characters).
-
-        Returns:
-            Full 32-character GUID.
-
-        Raises:
-            ValueError: If table is not in the allowed set.
-            ValueError: If partial is too short (< 8 chars).
-            ValueError: If no match found.
-            ValueError: If multiple matches found (ambiguous).
-        """
-        if table not in self._GUID_TABLES:
-            raise ValueError(f"Invalid table: {table}")
-        if len(partial) < 8:
-            raise ValueError(f"GUID prefix too short (minimum 8 chars): {partial}")
-
-        # Fast path: already a full GUID
-        if len(partial) == 32:
-            return partial
-
-        conn = sqlite3.connect(f"file:{self.book_path}?mode=ro", uri=True)
-        try:
-            rows = conn.execute(
-                f"SELECT guid FROM {table} WHERE guid LIKE ?",
-                (partial + "%",),
-            ).fetchall()
-        finally:
-            conn.close()
-
-        if len(rows) == 0:
-            raise ValueError(f"No {table[:-1]} found matching GUID prefix: {partial}")
-        if len(rows) > 1:
-            matches = [r[0] for r in rows]
-            raise ValueError(
-                f"Ambiguous GUID prefix '{partial}' matches {len(rows)} {table}: "
-                f"{', '.join(m[:12] + '...' for m in matches)}"
-            )
-        return rows[0][0]
-
-    @contextmanager
-    def open(
-        self, readonly: bool = True, max_retries: int = 3, retry_delay: float = 0.5
-    ) -> Generator[piecash.Book, None, None]:
-        """Context manager for book access with retry logic for locked files.
-
-        Args:
-            readonly: If True, open in read-only mode. Default True for safety.
-            max_retries: Number of retry attempts if file is locked. Default 3.
-            retry_delay: Seconds to wait between retries. Default 0.5.
-
-        Yields:
-            piecash.Book instance.
-
-        Raises:
-            GnuCashLockError: If the book is locked after all retries.
-            FileNotFoundError: If the book file doesn't exist.
-        """
-        last_error = None
-        for attempt in range(max_retries):
-            try:
-                start_time = time.time()
-                book = piecash.open_book(str(self.book_path), readonly=readonly, do_backup=False)
-                open_elapsed = (time.time() - start_time) * 1000
-                debug_logger.debug(
-                    f"Book opened (readonly={readonly}) in {open_elapsed:.0f}ms"
-                )
-                try:
-                    yield book
-                    return
-                finally:
-                    close_start = time.time()
-                    book.close()
-                    close_elapsed = (time.time() - close_start) * 1000
-                    debug_logger.debug(f"Book closed in {close_elapsed:.0f}ms")
-            except sqlite3.OperationalError as e:
-                last_error = e
-                error_msg = str(e).lower()
-                if "locked" in error_msg or "busy" in error_msg:
-                    if attempt < max_retries - 1:
-                        time.sleep(retry_delay * (attempt + 1))  # Exponential backoff
-                        continue
-                    raise GnuCashLockError(
-                        f"GnuCash book is locked (possibly by GnuCash or another process). "
-                        f"Close GnuCash and try again. Details: {e}"
-                    ) from e
-                raise  # Re-raise non-lock errors immediately
-
-        # Should not reach here, but just in case
-        if last_error:
-            raise last_error
-
-    def _find_account(self, book: piecash.Book, fullname: str) -> piecash.Account | None:
-        """Find an account by its full name path.
-
-        Args:
-            book: Open piecash book.
-            fullname: Full account path (e.g., 'Assets:Bank:Checking').
-
-        Returns:
-            Account if found, None otherwise.
-        """
-        for account in book.accounts:
-            if account.fullname == fullname:
-                return account
-        return None
-
-    def _find_transaction(
-        self, book: piecash.Book, guid: str
-    ) -> piecash.Transaction | None:
-        """Find a transaction by GUID or partial GUID prefix.
-
-        Args:
-            book: Open piecash book.
-            guid: Transaction GUID (full 32-char or 8+ char prefix).
-
-        Returns:
-            Transaction if found, None otherwise.
-
-        Raises:
-            ValueError: If partial GUID is ambiguous (matches multiple).
-        """
-        try:
-            full_guid = self._resolve_guid("transactions", guid)
-        except ValueError as e:
-            if "No transaction" in str(e):
-                return None
-            raise
-        for transaction in book.transactions:
-            if transaction.guid == full_guid:
-                return transaction
-        return None
-
-    @staticmethod
-    def _require_default_currency(book: piecash.Book) -> piecash.Commodity:
-        """Get the book's default currency, raising a clear error if missing.
-
-        Args:
-            book: Open piecash book.
-
-        Returns:
-            The default currency Commodity.
-
-        Raises:
-            ValueError: If the book has no default currency set.
-        """
-        dc = book.default_currency
-        if dc is None:
-            raise ValueError(
-                "Book has no default currency set. In GnuCash desktop, "
-                "set the default currency under Preferences > Accounts "
-                "(Edit menu on Linux/Windows, GnuCash menu on macOS), "
-                "or pass the currency parameter explicitly."
-            )
-        return dc
 
     def _find_commodity(
         self, book: piecash.Book, mnemonic: str, namespace: str = "CURRENCY"
@@ -532,29 +67,6 @@ class GnuCashBook:
             return book.commodities.get(mnemonic=mnemonic, namespace=namespace)
         except KeyError:
             return None
-
-    def _get_or_create_currency(
-        self, book: piecash.Book, mnemonic: str
-    ) -> piecash.Commodity:
-        """Get an existing currency or create it from ISO code.
-
-        Uses book.currencies which has a built-in fallback that auto-creates
-        currencies from the ISO 4217 table if they don't already exist.
-
-        Args:
-            book: Open piecash book.
-            mnemonic: ISO 4217 currency code (e.g., "USD", "EUR", "GBP").
-
-        Returns:
-            Commodity for the currency.
-
-        Raises:
-            ValueError: If mnemonic is not a valid ISO 4217 currency code.
-        """
-        try:
-            return book.currencies(mnemonic=mnemonic)
-        except (KeyError, ValueError) as e:
-            raise ValueError(f"Invalid currency code '{mnemonic}': {e}") from e
 
     def _find_budget(self, book: piecash.Book, name: str):
         """Find a budget by name.
@@ -705,14 +217,6 @@ class GnuCashBook:
             f"Invalid period: {period}. "
             f"Use None, 'all', an integer, or 'q1'-'q4'."
         )
-
-    def _collect_descendants(
-        self, account, result: set
-    ) -> None:
-        """Recursively collect all descendant accounts."""
-        for child in account.children:
-            result.add(child)
-            self._collect_descendants(child, result)
 
     # ============== Scheduled Transaction Helpers ==============
 
@@ -1499,7 +1003,7 @@ class GnuCashBook:
         all_names = frozenset(a.fullname for a in accounts)
         categorization = frozenset(
             a.fullname for a in accounts
-            if a.type not in GnuCashBook._FUNDING_ACCOUNT_TYPES
+            if a.type not in LegacyMixin._FUNDING_ACCOUNT_TYPES
         )
         return categorization if categorization else all_names
 
@@ -2856,31 +2360,6 @@ class GnuCashBook:
                 result["warnings"] = warnings
 
             return result
-
-    def _find_split(self, book: piecash.Book, guid: str) -> piecash.Split | None:
-        """Find a split by GUID or partial GUID prefix.
-
-        Args:
-            book: Open piecash book.
-            guid: Split GUID (full 32-char or 8+ char prefix).
-
-        Returns:
-            Split if found, None otherwise.
-
-        Raises:
-            ValueError: If partial GUID is ambiguous (matches multiple).
-        """
-        try:
-            full_guid = self._resolve_guid("splits", guid)
-        except ValueError as e:
-            if "No split" in str(e):
-                return None
-            raise
-        for transaction in book.transactions:
-            for split in transaction.splits:
-                if split.guid == full_guid:
-                    return split
-        return None
 
     # Valid reconcile states
     VALID_RECONCILE_STATES = {"n", "c", "y"}  # new, cleared, reconciled
@@ -5121,114 +4600,7 @@ class GnuCashBook:
                 "status": "closed",
             }
 
-    # ============== Account Slot Operations ==============
-
-    def get_account_slots(
-        self, account_name: str, key: str | None = None
-    ) -> dict:
-        """Read all slots (or a specific slot) from an account.
-
-        Args:
-            account_name: Full account path (e.g., "Liabilities:Credit Cards:Capital One").
-            key: Specific slot key to retrieve. If None, return all slots.
-
-        Returns:
-            Dict with account name and slots dict.
-
-        Raises:
-            ValueError: If account not found.
-        """
-        with self.open(readonly=True) as book:
-            account = self._find_account(book, account_name)
-            if not account:
-                raise ValueError(f"Account not found: {account_name}")
-
-            if key is not None:
-                try:
-                    value = account[key]
-                    slots = {key: str(value.value) if hasattr(value, 'value') else str(value)}
-                except KeyError:
-                    slots = {}
-            else:
-                slots = {}
-                for k, v in account.iteritems():
-                    slots[k] = str(v.value)
-
-            return {
-                "account": account_name,
-                "slots": slots,
-            }
-
-    def set_account_slot(
-        self, account_name: str, key: str, value: str
-    ) -> dict:
-        """Set a single key-value pair on an account.
-
-        Args:
-            account_name: Full account path.
-            key: Slot key (e.g., "apr", "credit_limit").
-            value: Slot value. Stored as string.
-
-        Returns:
-            Dict with account, key, value, and status ("created" or "updated").
-
-        Raises:
-            ValueError: If account not found.
-        """
-        with self.open(readonly=False) as book:
-            account = self._find_account(book, account_name)
-            if not account:
-                raise ValueError(f"Account not found: {account_name}")
-
-            # Check if key exists to determine created vs updated
-            existing = False
-            try:
-                account[key]
-                existing = True
-            except KeyError:
-                pass
-
-            account[key] = value
-            book.save()
-
-            return {
-                "account": account_name,
-                "key": key,
-                "value": value,
-                "status": "updated" if existing else "created",
-            }
-
-    def delete_account_slot(self, account_name: str, key: str) -> dict:
-        """Remove a slot from an account.
-
-        Args:
-            account_name: Full account path.
-            key: Slot key to remove.
-
-        Returns:
-            Dict with account, key, and status.
-
-        Raises:
-            ValueError: If account not found or key not found.
-        """
-        with self.open(readonly=False) as book:
-            account = self._find_account(book, account_name)
-            if not account:
-                raise ValueError(f"Account not found: {account_name}")
-
-            try:
-                account[key]
-            except KeyError:
-                raise ValueError(f"Slot key not found: {key}")
-
-            del account[key]
-            book.save()
-
-            return {
-                "account": account_name,
-                "key": key,
-                "status": "deleted",
-            }
+    # Account slot methods (get/set/delete) moved to book/admin.py as AdminMixin.
 
     # ------------------------------------------------------------------
     # Business: Customer / Vendor / Billterm
